+++ src/features/letter-list-senior/index.tsx
@@ -1,1 +1,2 @@
 export { default as SeniorLetterListFeature } from './containers/LetterListContainer'
+export { default as SeniorLetterArchiveListFeature } from './containers/LetterArchiveListContainer'
