+++ src/pages/latte-chat/letter/archive/senior/ui.tsx
@@ -1,0 +1,5 @@
+import { SeniorLetterArchiveListFeature } from '@/features/letter-list-senior'
+
+export default function SeniorLetterArchiveListPage() {
+  return <SeniorLetterArchiveListFeature />
+}
